+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/data/XYZColumnarData.kt
@@ -17,10 +17,10 @@ import space.kscience.kmath.structures.Buffer
 public interface XYZColumnarData<T, out X : T, out Y : T, out Z : T> : XYColumnarData<T, X, Y> {
     public val z: Buffer<Z>
 
-    override fun get(symbol: Symbol): Buffer<T> = when (symbol) {
+    override fun get(symbol: Symbol): Buffer<T>? = when (symbol) {
         Symbol.x -> x
         Symbol.y -> y
         Symbol.z -> z
-        else -> error("A column for symbol $symbol not found")
+        else -> null
     }
 }